+++ DotDoc.EntityFrameworkCore.Extensions/QueryExtensions.cs
@@ -2,13 +2,9 @@
 // This file is licensed to you under the MIT license.
 // See the License.txt file in the solution root for more information.
 
-using DotDoc.EntityFrameworkCore.Extensions.Constants;
 using DotDoc.EntityFrameworkCore.Extensions.Model;
-using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
-using Microsoft.EntityFrameworkCore.Storage;
 using System.Data;
-using System.Text.RegularExpressions;
 
 namespace DotDoc.EntityFrameworkCore.Extensions;
 
@@ -17,31 +13,6 @@ namespace DotDoc.EntityFrameworkCore.Extensions;
 /// </summary>
 public static class QueryExtensions
 {
-    #region Private Variables
-
-    /// <summary>
-    /// Split a SQL query into 2 groups:
-    ///     1: Before the ORDER BY clause.
-    ///     2: ORDER BY onwards.
-    /// </summary>
-    private static readonly Regex _sqlSplitOrderByRegex =
-        new (@"\A(.*)(\bORDER\s+BY\s+(?!.*?(?:\)|\s+)AS\s)(?:\((?>\((?<depth>)|\)(?<-depth>)|.?)*(?(depth)(?!))\)|[\[\]`""\w\(\)\.])+(?:\s+(?:ASC|DESC))?(?:\s*,\s*(?:\((?>\((?<depth>)|\)(?<-depth>)|.?)*(?(depth)(?!))\)|[\[\]`""\w\(\)\.])+(?:\s+(?:ASC|DESC))?)*)",
-            RegexOptions.RightToLeft | RegexOptions.IgnoreCase | RegexOptions.Multiline | RegexOptions.Singleline | RegexOptions.Compiled);
-
-    /// <summary>
-    /// Split a SQL query into 3 groups:
-    ///
-    ///     1:  SELECT
-    ///     2:  Field list
-    ///     3:  FROM onwards.
-    ///
-    /// </summary>
-    private static readonly Regex _sqlSplitColumnsRegex =
-        new (@"\A\s*(SELECT)\s+((?:\((?>\((?<depth>)|\)(?<-depth>)|.?)*(?(depth)(?!))\)|.)*?)(?<!,\s+)\b(FROM\b.+)",
-            RegexOptions.IgnoreCase | RegexOptions.Multiline | RegexOptions.Singleline | RegexOptions.Compiled);
-
-    #endregion Private Variables
-
     #region Public ExecuteScalar methods
 
     /// <summary>
@@ -52,7 +23,7 @@ public static class QueryExtensions
     /// <param name="sql">The interpolated string representing a SQL query with parameters.</param>
     /// <returns>The result of the query.</returns>
     public static T ExecuteScalarInterpolated<T>(this DatabaseFacade databaseFacade, FormattableString sql) =>
-        ExecuteScalar<T>(databaseFacade, sql.Format, sql.GetArguments());
+        QueryMethods.ExecuteScalar<T>(databaseFacade, sql.Format, sql.GetArguments());
 
     /// <summary>
     /// Executes a query with a single scalar result.
@@ -63,7 +34,7 @@ public static class QueryExtensions
     /// <param name="parameters">Parameters to use with the SQL.</param>
     /// <returns>The result of the query.</returns>
     public static T ExecuteScalarRaw<T>(this DatabaseFacade databaseFacade, string sql, params object[] parameters) =>
-        ExecuteScalar<T>(databaseFacade, sql, parameters);
+        QueryMethods.ExecuteScalar<T>(databaseFacade, sql, parameters);
 
     /// <summary>
     /// Executes a query with a single scalar result.
@@ -74,7 +45,7 @@ public static class QueryExtensions
     /// <param name="cancellationToken">A <see cref="CancellationToken"/> to observe while waiting for the task to complete.</param>
     /// <returns>The result of the query.</returns>
     public static Task<T> ExecuteScalarInterpolatedAsync<T>(this DatabaseFacade databaseFacade, FormattableString sql, CancellationToken cancellationToken = default) =>
-        ExecuteScalarAsync<T>(databaseFacade, sql.Format, sql.GetArguments(), cancellationToken);
+        QueryMethods.ExecuteScalarAsync<T>(databaseFacade, sql.Format, sql.GetArguments(), cancellationToken);
 
     /// <summary>
     /// Executes a query with a single scalar result.
@@ -86,7 +57,7 @@ public static class QueryExtensions
     /// <param name="parameters">Parameters to use with the SQL.</param>
     /// <returns>The result of the query.</returns>
     public static Task<T> ExecuteScalarRawAsync<T>(this DatabaseFacade databaseFacade, string sql, CancellationToken cancellationToken = default, params object[] parameters) =>
-        ExecuteScalarAsync<T>(databaseFacade, sql, parameters, cancellationToken);
+        QueryMethods.ExecuteScalarAsync<T>(databaseFacade, sql, parameters, cancellationToken);
 
     #endregion Public ExecuteScalar methods
 
@@ -99,7 +70,7 @@ public static class QueryExtensions
     /// <param name="sql">The interpolated string representing a SQL query with parameters.</param>
     /// <returns>A <see cref="DataTable"/> containing the results of the query.</returns>
     public static DataTable ExecuteQueryInterpolated(this DatabaseFacade databaseFacade, FormattableString sql) =>
-        ExecuteQuery(databaseFacade, sql.Format, sql.GetArguments());
+        QueryMethods.ExecuteQuery(databaseFacade, sql.Format, sql.GetArguments());
 
     /// <summary>
     /// Executes a query.
@@ -109,7 +80,7 @@ public static class QueryExtensions
     /// <param name="parameters">Parameters to use with the SQL.</param>
     /// <returns>A <see cref="DataTable"/> containing the results of the query.</returns>
     public static DataTable ExecuteQueryRaw(this DatabaseFacade databaseFacade, string sql, params object[] parameters) =>
-        ExecuteQuery(databaseFacade, sql, parameters);
+        QueryMethods.ExecuteQuery(databaseFacade, sql, parameters);
 
     /// <summary>
     /// Executes a query.
@@ -119,7 +90,7 @@ public static class QueryExtensions
     /// <param name="cancellationToken">A <see cref="CancellationToken"/> to observe while waiting for the task to complete.</param>
     /// <returns>A <see cref="DataTable"/> containing the results of the query.</returns>
     public static Task<DataTable> ExecuteQueryInterpolatedAsync(this DatabaseFacade databaseFacade, FormattableString sql, CancellationToken cancellationToken = default) =>
-        ExecuteQueryAsync(databaseFacade, sql.Format, sql.GetArguments(), cancellationToken);
+        QueryMethods.ExecuteQueryAsync(databaseFacade, sql.Format, sql.GetArguments(), cancellationToken);
 
     /// <summary>
     /// Executes a query.
@@ -130,7 +101,7 @@ public static class QueryExtensions
     /// <param name="parameters">Parameters to use with the SQL.</param>
     /// <returns>A <see cref="DataTable"/> containing the results of the query.</returns>
     public static Task<DataTable> ExecuteQueryRawAsync(this DatabaseFacade databaseFacade, string sql, CancellationToken cancellationToken = default, params object[] parameters) =>
-        ExecuteQueryAsync(databaseFacade, sql, parameters, cancellationToken);
+        QueryMethods.ExecuteQueryAsync(databaseFacade, sql, parameters, cancellationToken);
 
     #endregion Public ExecuteQuery methods
 
@@ -145,7 +116,7 @@ public static class QueryExtensions
     /// <param name="pageSize">Number of records per page.</param>
     /// <returns>A instance of <see cref="QueryPage"/> containing the page data (If the page number is past the end of the table then the it will become the last page).</returns>
     public static QueryPage ExecutePagedQueryInterpolated(this DatabaseFacade databaseFacade, FormattableString sql, long page, long pageSize) =>
-        ExecutePagedQuery(databaseFacade, sql.Format, sql.GetArguments(), page, pageSize);
+        QueryMethods.ExecutePagedQuery(databaseFacade, sql.Format, sql.GetArguments(), page, pageSize);
 
     /// <summary>
     /// Executes a query and returns the specified page of results.
@@ -157,7 +128,7 @@ public static class QueryExtensions
     /// <param name="parameters">Parameters to use with the SQL.</param>
     /// <returns>A instance of <see cref="QueryPage"/> containing the page data (If the page number is past the end of the table then the it will become the last page).</returns>
     public static QueryPage ExecutePagedQueryRaw(this DatabaseFacade databaseFacade, string sql, long page, long pageSize, params object[] parameters) =>
-        ExecutePagedQuery(databaseFacade, sql, parameters, page, pageSize);
+        QueryMethods.ExecutePagedQuery(databaseFacade, sql, parameters, page, pageSize);
 
     /// <summary>
     /// Executes a query and returns the specified page of results.
@@ -169,7 +140,7 @@ public static class QueryExtensions
     /// <param name="cancellationToken">A <see cref="CancellationToken"/> to observe while waiting for the task to complete.</param>
     /// <returns>A instance of <see cref="QueryPage"/> containing the page data (If the page number is past the end of the table then the it will become the last page).</returns>
     public static Task<QueryPage> ExecutePagedQueryInterpolatedAsync(this DatabaseFacade databaseFacade, FormattableString sql, long page, long pageSize, CancellationToken cancellationToken = default) =>
-        ExecutePagedQueryAsync(databaseFacade, sql.Format, sql.GetArguments(), page, pageSize, cancellationToken);
+        QueryMethods.ExecutePagedQueryAsync(databaseFacade, sql.Format, sql.GetArguments(), page, pageSize, cancellationToken);
 
     /// <summary>
     /// Executes a query and returns the specified page of results.
@@ -182,13 +153,11 @@ public static class QueryExtensions
     /// <param name="parameters">Parameters to use with the SQL.</param>
     /// <returns>A instance of <see cref="QueryPage"/> containing the page data (If the page number is past the end of the table then the it will become the last page).</returns>
     public static Task<QueryPage> ExecutePagedQueryRawAsync(this DatabaseFacade databaseFacade, string sql, long page, long pageSize, CancellationToken cancellationToken = default, params object[] parameters) =>
-        ExecutePagedQueryAsync(databaseFacade, sql, parameters, page, pageSize, cancellationToken);
+        QueryMethods.ExecutePagedQueryAsync(databaseFacade, sql, parameters, page, pageSize, cancellationToken);
 
     #endregion Public ExecutePagedQuery methods
 
     #region Public ExecuteNonQuery methods
-
-    // These are here for completeness.
 
     /// <summary>
     /// Executes a non query.
@@ -197,7 +166,7 @@ public static class QueryExtensions
     /// <param name="sql">The interpolated string representing a SQL query with parameters.</param>
     /// <returns>The number of rows affected.</returns>
     public static int ExecuteNonQueryInterpolated(this DatabaseFacade databaseFacade, FormattableString sql) =>
-        databaseFacade.ExecuteSqlInterpolated(sql);
+        QueryMethods.ExecuteNonQuery(databaseFacade, sql.Format, sql.GetArguments());
 
     /// <summary>
     /// Executes a non query.
@@ -207,7 +176,7 @@ public static class QueryExtensions
     /// <param name="parameters">Parameters to use with the SQL.</param>
     /// <returns>The number of rows affected.</returns>
     public static int ExecuteNonQueryRaw(this DatabaseFacade databaseFacade, string sql, params object[] parameters) =>
-        databaseFacade.ExecuteSqlRaw(sql, parameters);
+        QueryMethods.ExecuteNonQuery(databaseFacade, sql, parameters);
 
     /// <summary>
     /// Executes a non query.
@@ -217,7 +186,7 @@ public static class QueryExtensions
     /// <param name="cancellationToken">A <see cref="CancellationToken"/> to observe while waiting for the task to complete.</param>
     /// <returns>The number of rows affected.</returns>
     public static Task<int> ExecuteNonQueryInterpolatedAsync(this DatabaseFacade databaseFacade, FormattableString sql, CancellationToken cancellationToken = default) =>
-        databaseFacade.ExecuteSqlInterpolatedAsync(sql, cancellationToken);
+        QueryMethods.ExecuteNonQueryAsync(databaseFacade, sql.Format, sql.GetArguments(), cancellationToken);
 
     /// <summary>
     /// Executes a non query.
@@ -228,7 +197,7 @@ public static class QueryExtensions
     /// <param name="parameters">Parameters to use with the SQL.</param>
     /// <returns>The number of rows affected.</returns>
     public static Task<int> ExecuteNonQueryRawAsync(this DatabaseFacade databaseFacade, string sql, CancellationToken cancellationToken = default, params object[] parameters) =>
-        databaseFacade.ExecuteSqlRawAsync(sql, parameters, cancellationToken);
+        QueryMethods.ExecuteNonQueryAsync(databaseFacade, sql, parameters, cancellationToken);
 
     #endregion Public ExecuteNonQuery methods
 
@@ -241,7 +210,7 @@ public static class QueryExtensions
     /// <param name="sql">The interpolated string representing a SQL query with parameters.</param>
     /// <returns>The ID of the new record.</returns>
     public static long ExecuteInsertInterpolated(this DatabaseFacade databaseFacade, FormattableString sql) =>
-        ExecuteInsert(databaseFacade, sql.Format, sql.GetArguments());
+        QueryMethods.ExecuteInsert(databaseFacade, sql.Format, sql.GetArguments());
 
     /// <summary>
     /// Executes an insert command.
@@ -251,7 +220,7 @@ public static class QueryExtensions
     /// <param name="parameters">Parameters to use with the SQL.</param>
     /// <returns>The ID of the new record.</returns>
     public static long ExecuteInsertRaw(this DatabaseFacade databaseFacade, string sql, params object[] parameters) =>
-        ExecuteInsert(databaseFacade, sql, parameters);
+        QueryMethods.ExecuteInsert(databaseFacade, sql, parameters);
 
     /// <summary>
     /// Executes an insert command.
@@ -261,7 +230,7 @@ public static class QueryExtensions
     /// <param name="cancellationToken">A <see cref="CancellationToken"/> to observe while waiting for the task to complete.</param>
     /// <returns>The ID of the new record.</returns>
     public static Task<long> ExecuteInsertInterpolatedAsync(this DatabaseFacade databaseFacade, FormattableString sql, CancellationToken cancellationToken = default) =>
-        ExecuteInsertAsync(databaseFacade, sql.Format, sql.GetArguments(), cancellationToken);
+        QueryMethods.ExecuteInsertAsync(databaseFacade, sql.Format, sql.GetArguments(), cancellationToken);
 
     /// <summary>
     /// Executes an insert command.
@@ -272,306 +241,7 @@ public static class QueryExtensions
     /// <param name="parameters">Parameters to use with the SQL.</param>
     /// <returns>The ID of the new record.</returns>
     public static Task<long> ExecuteInsertRawAsync(this DatabaseFacade databaseFacade, string sql, CancellationToken cancellationToken = default, params object[] parameters) =>
-        ExecuteInsertAsync(databaseFacade, sql, parameters, cancellationToken);
+        QueryMethods.ExecuteInsertAsync(databaseFacade, sql, parameters, cancellationToken);
 
     #endregion Public ExecuteInsert methods
-
-    #region Private ExecuteScalar methods
-
-    /// <summary>
-    /// Executes a query with a single scalar result.
-    /// </summary>
-    /// <typeparam name="T">The type of result returned by the query.</typeparam>
-    /// <param name="databaseFacade">The <see cref="DatabaseFacade"/> for the context.</param>
-    /// <param name="sql">The SQL query to execute.</param>
-    /// <param name="parameters">Parameters to use with the SQL.</param>
-    /// <returns>The result of the query.</returns>
-    private static T ExecuteScalar<T>(DatabaseFacade databaseFacade, string sql, object[] parameters)
-    {
-        using ConcurrencyDetectorCriticalSectionDisposer criticalSectionDisposer = databaseFacade
-            .GetService<IConcurrencyDetector>()
-            .EnterCriticalSection();
-
-        RawSqlCommand rawSqlCommand = databaseFacade
-            .GetService<IRawSqlCommandBuilder>()
-            .Build(sql, parameters);
-
-        object value = rawSqlCommand
-            .RelationalCommand
-            .ExecuteScalar(new RelationalCommandParameterObject(databaseFacade.GetService<IRelationalConnection>(), rawSqlCommand.ParameterValues, null, null, null));
-
-        Type type = typeof(T);
-        return value == null ? default : (T)Convert.ChangeType(value, Nullable.GetUnderlyingType(type) ?? type);
-    }
-
-    /// <summary>
-    /// Executes a query with a single scalar result.
-    /// </summary>
-    /// <typeparam name="T">The type of result returned by the query.</typeparam>
-    /// <param name="databaseFacade">The <see cref="DatabaseFacade"/> for the context.</param>
-    /// <param name="sql">The SQL query to execute.</param>
-    /// <param name="parameters">Parameters to use with the SQL.</param>
-    /// <param name="cancellationToken">A <see cref="CancellationToken"/> to observe while waiting for the task to complete.</param>
-    /// <returns>The result of the query.</returns>
-    private static async Task<T> ExecuteScalarAsync<T>(DatabaseFacade databaseFacade, string sql, object[] parameters, CancellationToken cancellationToken)
-    {
-        using ConcurrencyDetectorCriticalSectionDisposer criticalSectionDisposer = databaseFacade
-            .GetService<IConcurrencyDetector>()
-            .EnterCriticalSection();
-
-        RawSqlCommand rawSqlCommand = databaseFacade
-            .GetService<IRawSqlCommandBuilder>()
-            .Build(sql, parameters);
-
-        object value = await rawSqlCommand
-            .RelationalCommand
-            .ExecuteScalarAsync(new RelationalCommandParameterObject(databaseFacade.GetService<IRelationalConnection>(), rawSqlCommand.ParameterValues, null, null, null), cancellationToken)
-            .ConfigureAwait(false);
-
-        Type type = typeof(T);
-        return value == null ? default : (T)Convert.ChangeType(value, Nullable.GetUnderlyingType(type) ?? type);
-    }
-
-    #endregion Private ExecuteScalar methods
-
-    #region Private ExecuteQuery methods
-
-    /// <summary>
-    /// Executes a query.
-    /// </summary>
-    /// <param name="databaseFacade">The <see cref="DatabaseFacade"/> for the context.</param>
-    /// <param name="sql">The SQL query to execute.</param>
-    /// <param name="parameters">Parameters to use with the SQL.</param>
-    /// <returns>A <see cref="DataTable"/> containing the results of the query.</returns>
-    private static DataTable ExecuteQuery(DatabaseFacade databaseFacade, string sql, object[] parameters)
-    {
-        using ConcurrencyDetectorCriticalSectionDisposer criticalSectionDisposer = databaseFacade
-            .GetService<IConcurrencyDetector>()
-            .EnterCriticalSection();
-
-        RawSqlCommand rawSqlCommand = databaseFacade
-            .GetService<IRawSqlCommandBuilder>()
-            .Build(sql, parameters);
-
-        using RelationalDataReader relationalDataReader = rawSqlCommand
-            .RelationalCommand
-            .ExecuteReader(new RelationalCommandParameterObject(databaseFacade.GetService<IRelationalConnection>(), rawSqlCommand.ParameterValues, null, null, null));
-
-        DataTable dataTable = new ();
-        dataTable.Load(relationalDataReader.DbDataReader);
-        return dataTable;
-    }
-
-    /// <summary>
-    /// Executes a query.
-    /// </summary>
-    /// <param name="databaseFacade">The <see cref="DatabaseFacade"/> for the context.</param>
-    /// <param name="sql">The SQL query to execute.</param>
-    /// <param name="parameters">Parameters to use with the SQL.</param>
-    /// <param name="cancellationToken">A <see cref="CancellationToken"/> to observe while waiting for the task to complete.</param>
-    /// <returns>A <see cref="DataTable"/> containing the results of the query.</returns>
-    private static async Task<DataTable> ExecuteQueryAsync(DatabaseFacade databaseFacade, string sql, object[] parameters, CancellationToken cancellationToken)
-    {
-        using ConcurrencyDetectorCriticalSectionDisposer criticalSectionDisposer = databaseFacade
-            .GetService<IConcurrencyDetector>()
-            .EnterCriticalSection();
-
-        RawSqlCommand rawSqlCommand = databaseFacade
-            .GetService<IRawSqlCommandBuilder>()
-            .Build(sql, parameters);
-
-        await using RelationalDataReader relationalDataReader = await rawSqlCommand
-            .RelationalCommand
-            .ExecuteReaderAsync(new RelationalCommandParameterObject(databaseFacade.GetService<IRelationalConnection>(), rawSqlCommand.ParameterValues, null, null, null), cancellationToken)
-            .ConfigureAwait(false);
-
-        DataTable dataTable = new ();
-        dataTable.Load(relationalDataReader.DbDataReader);
-        return dataTable;
-    }
-
-    #endregion Private ExecuteQuery methods
-
-    #region Private ExecutePagedQuery methods
-
-    /// <summary>
-    /// Executes a query and returns the specified page of results.
-    /// </summary>
-    /// <param name="databaseFacade">The <see cref="DatabaseFacade"/> for the context.</param>
-    /// <param name="sql">The SQL query to execute.</param>
-    /// <param name="parameters">Parameters to use with the SQL.</param>
-    /// <param name="page">Page number to return (starting at 0).</param>
-    /// <param name="pageSize">Number of records per page.</param>
-    /// <returns>A instance of <see cref="QueryPage"/> containing the page data (If the page number is past the end of the table then the it will become the last page).</returns>
-    private static QueryPage ExecutePagedQuery(DatabaseFacade databaseFacade, string sql, object[] parameters, long page, long pageSize)
-    {
-        string countSql = ConvertQueryToCount(sql);
-        long recordCount;
-        long pageCount;
-        DataTable dataTable;
-
-        do
-        {
-            recordCount = ExecuteScalar<long>(databaseFacade, countSql, parameters);
-
-            pageCount = (recordCount + pageSize - 1) / pageSize;
-            if (page >= pageCount)
-            {
-                page = pageCount > 0 ? pageCount - 1 : 0;
-            }
-
-            (string pageSql, object[] pageParameters) = LimitQuery(databaseFacade, sql, page, pageSize, parameters);
-
-            dataTable = ExecuteQuery(databaseFacade, pageSql, pageParameters);
-
-        } while (dataTable.Rows.Count == 0 && page > 0);
-
-        return new QueryPage(page, pageSize, recordCount, pageCount, dataTable);
-    }
-
-    /// <summary>
-    /// Executes a query and returns the specified page of results.
-    /// </summary>
-    /// <param name="databaseFacade">The <see cref="DatabaseFacade"/> for the context.</param>
-    /// <param name="sql">The SQL query to execute.</param>
-    /// <param name="parameters">Parameters to use with the SQL.</param>
-    /// <param name="page">Page number to return (starting at 0).</param>
-    /// <param name="pageSize">Number of records per page.</param>
-    /// <param name="cancellationToken">A <see cref="CancellationToken"/> to observe while waiting for the task to complete.</param>
-    /// <returns>A instance of <see cref="QueryPage"/> containing the page data (If the page number is past the end of the table then the it will become the last page).</returns>
-    private static async Task<QueryPage> ExecutePagedQueryAsync(DatabaseFacade databaseFacade, string sql, object[] parameters, long page, long pageSize, CancellationToken cancellationToken)
-    {
-        string countSql = ConvertQueryToCount(sql);
-        long recordCount;
-        long pageCount;
-        DataTable dataTable;
-
-        do
-        {
-            recordCount = await ExecuteScalarAsync<long>(databaseFacade, countSql, parameters, cancellationToken).ConfigureAwait(false);
-
-            pageCount = (recordCount + pageSize - 1) / pageSize;
-            if (page >= pageCount)
-            {
-                page = pageCount > 0 ? pageCount - 1 : 0;
-            }
-
-            (string pageSql, object[] pageParameters) = LimitQuery(databaseFacade, sql, page, pageSize, parameters);
-
-            dataTable = await ExecuteQueryAsync(databaseFacade, pageSql, pageParameters, cancellationToken).ConfigureAwait(false);
-
-        } while (dataTable.Rows.Count == 0 && page > 0);
-
-        return new QueryPage(page, pageSize, recordCount, pageCount, dataTable);
-    }
-
-    #endregion Private ExecutePagedQuery methods
-
-    #region Private ExecuteInsert methods
-
-    /// <summary>
-    /// Executes an insert command.
-    /// </summary>
-    /// <param name="databaseFacade">The <see cref="DatabaseFacade"/> for the context.</param>
-    /// <param name="sql">The SQL query to execute.</param>
-    /// <param name="parameters">Parameters to use with the SQL.</param>
-    /// <returns>The ID of the new record.</returns>
-    private static long ExecuteInsert(DatabaseFacade databaseFacade, string sql, object[] parameters) =>
-        ExecuteScalar<long>(databaseFacade, GetLastInsertId(databaseFacade, sql), parameters);
-
-    /// <summary>
-    /// Executes an insert command.
-    /// </summary>
-    /// <param name="databaseFacade">The <see cref="DatabaseFacade"/> for the context.</param>
-    /// <param name="sql">The SQL query to execute.</param>
-    /// <param name="parameters">Parameters to use with the SQL.</param>
-    /// <param name="cancellationToken">A <see cref="CancellationToken"/> to observe while waiting for the task to complete.</param>
-    /// <returns>The ID of the new record.</returns>
-    private static Task<long> ExecuteInsertAsync(DatabaseFacade databaseFacade, string sql, object[] parameters, CancellationToken cancellationToken) =>
-        ExecuteScalarAsync<long>(databaseFacade, GetLastInsertId(databaseFacade, sql), parameters, cancellationToken);
-
-    #endregion Private ExecuteInsert methods
-
-    #region Private Utility methods
-
-    /// <summary>
-    /// Convert a SQL query into a query to count the number of records.
-    /// </summary>
-    /// <param name="sql">The SQL query to execute.</param>
-    /// <returns>The SQL query with the field list changed to COUNT(*).</returns>
-    private static string ConvertQueryToCount(string sql)
-    {
-        // First remove the Order By clause if there is one.
-        // These are meaningless in this situation as we are fetching a count of records.
-        Match match = _sqlSplitOrderByRegex.Match(sql);
-        if (match.Success)
-        {
-            sql = match.Groups[1].Value;
-        }
-
-        // Next replace the selected columns with Count(*).
-        match = _sqlSplitColumnsRegex.Match(sql);
-        if (!match.Success)
-        {
-            throw new InvalidOperationException("Unable to split query.");
-        }
-
-        return $"{match.Groups[1]} COUNT(*) {match.Groups[3]}";
-    }
-
-    /// <summary>
-    /// Add a limit / offset clause to a query.
-    /// </summary>
-    /// <param name="databaseFacade">The <see cref="DatabaseFacade"/> for the context.</param>
-    /// <param name="sql">The SQL query to execute.</param>
-    /// <param name="page">Page number to return (starting at 0).</param>
-    /// <param name="pageSize">Number of records per page.</param>
-    /// <param name="parameters">Parameters to use with the SQL.</param>
-    /// <returns>A tuple containting the modified SQL and parameters array.</returns>
-    private static (string, object[]) LimitQuery(DatabaseFacade databaseFacade, string sql, long page, long pageSize, object[] parameters)
-    {
-        switch (databaseFacade.GetDatabaseType())
-        {
-            case DatabaseType.Sqlite:
-                // Sqlite just needs "LIMIT x OFFSET y" at the end of a query
-                parameters = parameters.Concat(new object[] { pageSize, page * pageSize }).ToArray();
-
-                sql = $"{sql} LIMIT {{{parameters.Length - 2}}} OFFSET {{{parameters.Length - 1}}}";
-                break;
-
-            case DatabaseType.SqlServer:
-                // SQL Server must have "OFFSET x ROWS FETCH NEXT y ROWS ONLY" following the final "ORDER BY" clause.
-                // If there is no "Order By" then add a dummy one "Order By
-                parameters = parameters.Concat(new object[] { pageSize, page * pageSize }).ToArray();
-
-                Match match = _sqlSplitOrderByRegex.Match(sql);
-                sql = match.Success
-                    ? $"{match.Groups[1]} {match.Groups[2]}"
-                    : $"{sql} ORDER BY (SELECT NULL)";
-
-                sql += $" OFFSET {{{parameters.Length - 1}}} ROWS FETCH NEXT {{{parameters.Length - 2}}} ROWS ONLY";
-                break;
-
-            default:
-                throw new InvalidOperationException("Unsupported database type");
-        }
-
-        return (sql, parameters);
-    }
-
-    /// <summary>
-    /// Add a sql query to get the last created ID to a query.
-    /// </summary>
-    /// <param name="databaseFacade">The <see cref="DatabaseFacade"/> for the context.</param>
-    /// <param name="sql">The SQL query to execute.</param>
-    /// <returns>The existing sql query with the query to fetch the last ID attached.</returns>
-    private static string GetLastInsertId(DatabaseFacade databaseFacade, string sql) =>
-        databaseFacade.GetDatabaseType() switch
-        {
-            DatabaseType.Sqlite => $"{sql}; SELECT LAST_INSERT_ROWID();",
-            DatabaseType.SqlServer => $"{sql}; SELECT SCOPE_IDENTITY();",
-            _ => throw new InvalidOperationException("Unsupported database type"),
-        };
-
-    #endregion Private Utility methods
 }